refactor(queue): add WORKER_NUM and QUEUE_SIZE
