+++ docs/config/01-config.docker.ini
@@ -46,7 +46,8 @@ REDIS_HOST = redis:6379
 REDIS_TIMEOUT_MILLI_TS = 10
 
 [go-openbbsmiddleware:queue]
-N_COMMENT_QUEUE = 4096
+WORKER_NUM = 4
+QUEUE_SIZE = 4096
 
 [go-pttbbs:api] # needs to be aligned with go-pttbbs
 JWT_SECRET = jwt_secret # JWT secret
