+++ docs/config/03-config.local-docker.ini
@@ -40,7 +40,8 @@ REDIS_HOST = redis:6379
 REDIS_TIMEOUT_MILLI_TS = 10
 
 [go-openbbsmiddleware:queue]
-N_COMMENT_QUEUE = 4096
+WORKER_NUM = 4
+QUEUE_SIZE = 4096
 
 [go-pttbbs:api] # needs to be aligned with go-pttbbs
 JWT_SECRET = jwt_secret # JWT secret
