--- queue/testinit.go
@@ -1,15 +1,0 @@
-package queue
-
-var IsTest = false
-
-func SetIsTest() {
-	IsTest = true
-
-	Start()
-}
-
-func UnsetIsTest() {
-	IsTest = false
-
-	Close()
-}
